refactor(configs): :recycle: repackage abstract config
