+++ src/fastapi_factory_utilities/core/app/base/__init__.py
@@ -1,7 +1,6 @@
 """Package for the base application, abstract config classes and related exceptions."""
 
 from .application import BaseApplication
-from .config_abstract import AppConfigAbstract
 from .exceptions import (
     ApplicationConfigFactoryException,
     ApplicationFactoryException,
@@ -10,7 +9,6 @@ from .exceptions import (
 
 __all__: list[str] = [
     "BaseApplication",
-    "AppConfigAbstract",
     "ApplicationConfigFactoryException",
     "ApplicationFactoryException",
     "BaseApplicationException",
